shell redirection in Project#build replaced with two seperate calls to execute rake and write build_log
As with quite a few network filesystems, DuoStack's Cloud Storage mount didn't like out little shell redirection from rake to build_log. In order to retrieve the output as well as the exitcode of the rake invocation, I handwrote a case of Environment#system to yield both to a block where build status processing is now done. While there are no tests around that method (and I didn't have the patience to figure out how to write tests for shell escapes!) the method still maintails the contract of returning the invocation's exitcode.